# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "DiagnosisReportResponse": {
        "properties": {
          "confidence": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "recommendations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "root_causes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Root Causes",
            "type": "array"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "workflow_run_id": {
            "format": "uuid",
            "title": "Workflow Run Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workflow_run_id",
          "version",
          "schema_version",
          "summary",
          "root_causes",
          "evidence",
          "recommendations",
          "confidence",
          "model_name",
          "prompt_version",
          "created_at"
        ],
        "title": "DiagnosisReportResponse",
        "type": "object"
      },
...
      "HumanDecisionAction": {
        "enum": [
          "approve",
          "reject",
          "reanalyze"
        ],
        "title": "HumanDecisionAction",
        "type": "string"
      },
...
      "HumanDecisionResponse": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/HumanDecisionAction"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "diagnosis_report_id": {
            "format": "uuid",
            "title": "Diagnosis Report Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "diagnosis_report_id",
          "idempotency_key",
          "action",
          "comment",
          "actor",
          "created_at"
        ],
        "title": "HumanDecisionResponse",
        "type": "object"
      },
...
      "WorkflowAcceptedResponse": {
        "properties": {
          "dispatched": {
            "title": "Dispatched",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowRunStatus"
          },
          "workflow_run_id": {
            "format": "uuid",
            "title": "Workflow Run Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_run_id",
          "status",
          "dispatched"
        ],
        "title": "WorkflowAcceptedResponse",
        "type": "object"
      },
      "WorkflowDetailResponse": {
        "properties": {
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HumanDecisionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiagnosisReportResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "run": {
            "$ref": "#/components/schemas/WorkflowRunResponse"
          }
        },
        "required": [
          "run"
        ],
        "title": "WorkflowDetailResponse",
        "type": "object"
      },
...
      "WorkflowRunResponse": {
        "properties": {
          "alert_id": {
            "format": "uuid",
            "title": "Alert Id",
            "type": "string"
          },
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Node"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowRunStatus"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_version": {
            "title": "Workflow Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "alert_id",
          "thread_id",
          "workflow_version",
          "status",
          "current_node",
          "attempt",
          "error_code",
          "error_message",
          "started_at",
          "finished_at",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowRunResponse",
        "type": "object"
      },
      "WorkflowRunStatus": {
        "enum": [
          "queued",
          "running",
          "waiting_for_approval",
          "reanalyzing",
          "completed",
          "rejected",
          "failed"
        ],
        "title": "WorkflowRunStatus",
        "type": "string"
      },
      "WorkflowStartCommand": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "maxLength": 160,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key"
        ],
        "title": "WorkflowStartCommand",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Alert Sage API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/alerts/{alert_id}/workflow": {
      "get": {
        "operationId": "get_workflow_api_v1_alerts__alert_id__workflow_get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Workflow",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "start_workflow_api_v1_alerts__alert_id__workflow_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStartCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}